Make it possible to simulate square antennas with more than 64 elements

+++ src/mmwave/model/antenna-array-model.h
@@ -58,8 +58,8 @@ public:
   void ChangeToOmniTx ();
   bool IsOmniTx ();
   double GetRadiationPattern (double vangle, double hangle = 0);
-  Vector GetAntennaLocation (uint8_t index, uint8_t* antennaNum);
-  void SetSector (uint8_t sector, uint8_t *antennaNum, double elevation = 90);
+  Vector GetAntennaLocation (uint16_t index, uint16_t* antennaNum);
+  void SetSector (uint8_t sector, uint16_t *antennaNum, double elevation = 90);
 
   void SetPlanesNumber (double planesNumber);
   double GetPlanesId (void);
